Version v. 2.0.0 [NEW RELEASE - v. 2.0.0]:
- New feature: Automatic algorithm switching for large datasets
For datasets with 10,000 molecules or more, MASSA now (automatically) switches from Hierarchical Clustering Analysis (HCA) to MiniBatch K-Means for improved scalability and performance. This behavior can be controlled with the new flag: -a / --large-datasets. Dendrogram plotting is disabled in this mode, and the -f / --dendrogram-plot flag is automatically ignored.
- New feature: MASSA can now be called from Python scripts using MASSA_Algorithm.pyMASSA.py_massa().
The feature is already available, but instructions for using it will be provided in upcoming updates/versions.
- Bug fix: Fixed "NaN/None values in y" error.
This error, caused by duplicate molecule names, is now resolved. Molecules with identical names are automatically renamed by appending their original index to ensure uniqueness.
- Change: MASSA now ensures preservation of original input order, but be aware that molecules with chemical errors are now skipped by default during processing and not included in the output.